Added registration success page
Added a registration success page that the routes the user back to the
home page. I also added a note to the home page regarding the clicking
of the Borrow My Angel icon being used as a button to return back to the
home screen. This will obviously not be a feature once the user is
logged into the client services.