Registration encoded for back end form
Changed the registrations to allow for submission of the form data in
the requested encoded data type. Furthermore, I created a basic HUB
component that would be used once someone logs into the system through
the log in page. It is currently empty; however, there are drawings on
the whiteboard with a few sketches of what we would like the finalized
HUB to look at. Thanks. :)